reworking hash
