+++ client/package.json
@@ -7,7 +7,6 @@
     "@types/react": "^16.9.23",
     "ace-builds": "^1.4.8",
     "cra-template": "1.0.0",
-    "crypto": "^1.0.1",
     "prismjs": "^1.19.0",
     "react": "^16.12.0",
     "react-ace": "^8.0.0",
